fix(agreement search): refector agreement search code to decouple search in wizard from search as a tool (#4407)
