--- packages/code-du-travail-frontend/src/outils/ConventionCollective/common/EmbeddedInForm.tsx
@@ -1,24 +1,0 @@
-import React, { FormEventHandler, ForwardedRef } from "react";
-
-const EmbeddedInForm = ({
-  enable,
-  reference,
-  onSubmit,
-  children,
-}: {
-  enable: boolean;
-  reference: ForwardedRef<HTMLFormElement>;
-  onSubmit: FormEventHandler;
-  children: React.ReactNode;
-}): JSX.Element => {
-  if (enable) {
-    return (
-      <form ref={reference} onSubmit={onSubmit}>
-        {children}
-      </form>
-    );
-  }
-  return <>{children}</>;
-};
-
-export default EmbeddedInForm;
