+++ packages/code-du-travail-frontend/src/outils/common/Agreement/AgreementSearch/AgreementSearch.tsx
@@ -1,17 +1,23 @@
-import { theme, Toast } from "@socialgouv/cdtn-ui";
+import {
+  Paragraph,
+  Section as SectionUi,
+  theme,
+  Toast,
+} from "@socialgouv/cdtn-ui";
 import React from "react";
 import { Field } from "react-final-form";
 import styled from "styled-components";
 
 import { Agreement } from "../../../../conventions/Search/api/type";
-import { AgreementSearchStep } from "../../../ConventionCollective/steps/AgreementSearch";
 import { TrackingProps } from "../../../ConventionCollective/types";
 import { ErrorField } from "../../ErrorField";
 import { Question } from "../../Question";
 import { required } from "../../validators";
 import { AGREEMENT_NAME } from "../form-constants";
-import ShowAlert from "../ShowAlert";
+import ShowAlert from "../components/ShowAlert";
 import { AgreementSupportInfo } from "../types";
+import { renderResults } from "./AgreementSearchResult";
+import { SearchAgreementInput } from "./AgreementInput/SearchAgreementInput";
 
 export type Props = {
   supportedAgreements: AgreementSupportInfo[];
@@ -51,11 +57,16 @@ const AgreementSearch = ({
     );
   }
   return (
-    <>
-      <AgreementSearchStep
+    <Section>
+      <Paragraph noMargin fontWeight="600" fontSize="default">
+        Précisez et sélectionnez votre convention collective
+      </Paragraph>
+      <SearchAgreementInput
         onUserAction={onUserAction}
-        embeddedForm={false}
-        onSelectAgreement={onSelectAgreement}
+        renderResults={renderResults({
+          onSelectAgreement,
+          onUserAction,
+        })}
       />
       <ErrorField
         name={AGREEMENT_NAME}
@@ -70,12 +81,18 @@ const AgreementSearch = ({
           return <input {...props} {...input} />;
         }}
       />
-    </>
+    </Section>
   );
 };
 
 const { spacings } = theme;
 
+const Section = styled(SectionUi)`
+  padding-top: 0;
+  label {
+    font-weight: 400;
+  }
+`;
 const SelectedAgreement = styled(Toast)`
   margin-bottom: ${spacings.base};
 `;
